Fix conversion warnings for the Graphics module
- Fix conversion and shadowing warnings
- Add SYSTEM indicator for stb_*, FreeType and other headers

+++ src/SFML/Graphics/Color.cpp
@@ -69,7 +69,7 @@ a(alpha)
 
 ////////////////////////////////////////////////////////////
 Color::Color(Uint32 color) :
-r((color & 0xff000000) >> 24),
+r(static_cast<Uint8>((color & 0xff000000) >> 24)),
 g(static_cast<Uint8>((color & 0x00ff0000) >> 16)),
 b((color & 0x0000ff00) >> 8 ),
 a((color & 0x000000ff) >> 0 )
@@ -81,7 +81,7 @@ a((color & 0x000000ff) >> 0 )
 ////////////////////////////////////////////////////////////
 Uint32 Color::toInteger() const
 {
-    return (r << 24) | (g << 16) | (b << 8) | a;
+    return static_cast<Uint32>((r << 24) | (g << 16) | (b << 8) | a);
 }
